+++ modules/speaker.py
@@ -5,19 +5,30 @@
 
 """
 
-from typing import NoReturn
+from typing import NoReturn, Union
 
-import pyttsx3
-
+from modules.exceptions import InvalidEnvVars
 from modules.logger import logger
-from modules.models import settings
+from modules.models import audio_driver, env, settings
 
-audio_driver = pyttsx3.init()
-voices = audio_driver.getProperty("voices")  # gets the list of voices available
-voice_model = "Daniel" if settings.macos else "David"
-for ind_d, voice_id in enumerate(voices):  # noqa
-    if voice_id.name == voice_model or voice_model in voice_id.name:
-        audio_driver.setProperty("voice", voices[ind_d].id)  # noqa
+voices: Union[list, object] = audio_driver.getProperty("voices")  # gets the list of voices available
+voice_names = [__voice.name for __voice in voices]
+if not env.voice_name:
+    if settings.os == "Darwin":
+        env.voice_name = "Daniel"
+    elif settings.os == "Windows":
+        env.voice_name = "David"
+    elif settings.os == "Linux":
+        env.voice_name = "english-us"
+elif env.voice_name not in voice_names:
+    raise InvalidEnvVars(
+        f"{env.voice_name!r} is not available.\nAvailable voices are: {', '.join(voice_names)}"
+    )
+for ind_d, voice in enumerate(voices):  # noqa
+    if voice.name == env.voice_name:
+        logger.debug(voice.__dict__)
+        audio_driver.setProperty("voice", voices[ind_d].id)
+        audio_driver.setProperty("rate", env.voice_rate)
         break
 else:
     logger.info("Using default voice model.")